Gear Up: Integrated latest dev.
[git-p4: depot-paths = "//depot/traktor/": change = 14486]

+++ code/Input/X11/MouseDeviceX11.h
@@ -60,7 +60,7 @@ private:
 	int m_deviceId;
 	bool m_connected;
 	bool m_exclusive;
-	float m_raw[3];
+	int32_t m_raw[3];
 	float m_axis[3];
 	int32_t m_position[2];
 	int32_t m_warped[2];
